Allow configurations to be passed into BackendServer
The `BackendServer` now takes an option called `:configurations` which
is a list of known `WiFiConfiguration.configurations()`. This allows us
to pass previously configured networks into the `BackendServer`.

The two benefits to this. First, if a user previously configured their
device and then ran the wizard now load the configuration on start. This
fixes an issue of the wizard clearing previous configurations. Secondly,
we the user is presented with the list of their configurations in the UI
the old configure is there. This will allow them to edit (say update the
password if it has changed since they last configured) or delete the
configuration.